add the ability to scan OR dependencies
