+++ mobros/dependency_manager/conflict_solver.py
@@ -2,171 +2,9 @@
 import sys
 
 import mobros.utils.logger as logging
-from mobros.exceptions import InstallCandidateNotFoundException
 from mobros.utils import apt_utils, version_utils
 
 
-def find_lowest_top_rule(version_rules):
-    """Function to find the lowest 'lower than' rule of a dependency.
-
-    Args:
-        version_rules (list): list of all version rules from the workspace
-
-    Returns:
-        version_rule: the lowest 'lower than' rule within the list or None if no 'lower than' rule found
-    """
-    included = True
-    bottom_limit = []
-
-    for rule in version_rules:
-        if "version_lt" == rule["operator"]:
-            bottom_limit.append(
-                {
-                    "version": rule["version"],
-                    "included": not included,
-                    "from": rule["from"],
-                }
-            )
-        elif "version_lte" == rule["operator"]:
-            bottom_limit.append(
-                {"version": rule["version"], "included": included, "from": rule["from"]}
-            )
-
-    if len(bottom_limit) == 0:
-        return None
-
-    version_utils.order_rule_versions(bottom_limit)
-    return bottom_limit[0]
-
-
-def find_highest_bottom_rule(version_rules):
-    """Function to find the highest 'greater than' rule of a dependency.
-
-    Args:
-        version_rules (list): list of all version rules from the workspace
-
-    Returns:
-        version_rule: the highest 'greater than' rule within the list or None if no 'greater than' rule found
-    """
-    included = True
-    top_limit = []
-
-    for rule in version_rules:
-        if "version_gt" == rule["operator"]:
-            top_limit.append(
-                {
-                    "version": rule["version"],
-                    "included": not included,
-                    "from": rule["from"],
-                }
-            )
-        elif "version_gte" == rule["operator"]:
-            top_limit.append(
-                {"version": rule["version"], "included": included, "from": rule["from"]}
-            )
-
-    if len(top_limit) == 0:
-        return None
-
-    version_utils.order_rule_versions(top_limit, reverse=True)
-    return top_limit[0]
-
-
-def find_candidate_online(deb_name, version_rules):
-    """Function that is able to find the candidate version from the apt cache that passes
-    the version rules from all dependencies in the workspace
-
-    Args:
-        deb_name (str): debian package name
-        version_rules (list): list of all version rules from the workspace
-
-    Returns:
-        version: The candidate version to be installed from apt of a given debian package
-    """
-    avaiable_versions = apt_utils.get_package_avaiable_versions(deb_name)
-
-    if not avaiable_versions:
-        msg = "Unable to find online versions of package " + deb_name + "\n"
-        msg += "Tip: Check if mobros was able to update your apt cache (apt update)! Either run mobros with sudo or execute 'apt update' beforehand"
-        raise InstallCandidateNotFoundException(msg)
-
-    found, equals_rule = version_utils.find_equals_rule(version_rules)
-    if found:
-        if equals_rule["version"] not in avaiable_versions:
-            msg = "Unable to find a candidate for " + deb_name + "\n"
-            msg += (
-                "Filter rule: "
-                + equals_rule["operator"]
-                + " ("
-                + equals_rule["version"]
-                + " ) from "
-                + equals_rule["from"]
-                + "\n"
-            )
-            msg += "Avaiable online: " + str(avaiable_versions)
-            raise InstallCandidateNotFoundException(msg)
-
-        return [equals_rule["version"]]
-
-    top_limit_rule = find_lowest_top_rule(version_rules)
-    bottom_limit_rule = find_highest_bottom_rule(version_rules)
-
-    remaining_versions = avaiable_versions
-    top_rule_message = bottom_rule_message = "any"
-    if top_limit_rule:
-        remaining_versions = version_utils.filter_through_top_rule(
-            avaiable_versions, top_limit_rule, deb_name
-        )
-        top_rule_message = "<"
-        if top_limit_rule["included"]:
-            top_rule_message += "="
-
-        top_rule_message += (
-            " (" + top_limit_rule["version"] + ") from " + top_limit_rule["from"]
-        )
-
-    if bottom_limit_rule:
-        remaining_versions = version_utils.filter_through_bottom_rule(
-            remaining_versions, bottom_limit_rule, deb_name
-        )
-        bottom_rule_message = ">"
-        if bottom_limit_rule["included"]:
-            bottom_rule_message += "="
-        bottom_rule_message += (
-            " (" + bottom_limit_rule["version"] + ") from " + bottom_limit_rule["from"]
-        )
-
-    if len(remaining_versions) == 1:
-        logging.debug(
-            "[Find candidates online] Final decision for "
-            + deb_name
-            + " is: "
-            + str(remaining_versions[0])
-        )
-        return remaining_versions
-
-    if not remaining_versions:
-        msg = "Unable to find a candidate for " + deb_name + "\n"
-        msg += "Top limit rule: " + top_rule_message + "\n"
-        msg += "Bottom limit rule: " + bottom_rule_message + "\n"
-        msg += "From the Avaiable online: " + str(avaiable_versions)
-        raise InstallCandidateNotFoundException(msg)
-
-    logging.debug("-------------------------------------------------")
-    logging.debug(
-        "[Find candidates online] Pkg:"
-        + deb_name
-        + ". After top and bottom filters, remaining versions"
-        + str(remaining_versions)
-    )
-    logging.debug("-------------------------------------------------")
-    logging.debug(
-        "[Find candidates online] Pkg:"
-        + deb_name
-        + ". Final decision is: "
-        + str(remaining_versions[0])
-    )
-    return remaining_versions
 
 
 # def check_deps(name, version, ideb_name, ideb_version, visited=None):
